Move 'App.jsx' back, as 'App.js' to src folder
